# openapi.yaml
openapi: 3.0.0
info:
  title: Personal System
  version: 1.0.0
  description: >-
    A system that allows the easyer management of all the company personal. This
    API returns a JWT token in the /auth/login endpoint tha needs to be decoded
    to determine the user interface, which is different from the employee
    interface and the human resources interface
servers:
  - url: http://localhost:8080/api
    description: Generated server url
paths:
...
  /monthly-payments:
    get:
      tags:
        - monthly-payment-controller
      operationId: getPaymentsByPersonal
      summary: acces to human resources and personal that request his own information
      security:
        - JWTAuth:
            - human resources
            - employee
      parameters:
        - name: personalId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: month
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/MonthlyPaymentDto'
                  - type: object
        '404':
          description: not found
          content:
            application/text*:
              - schema: null
    post:
      tags:
        - monthly-payment-controller
      operationId: generatePayment
      summary: acces to human resources
      security:
        JWTAuth:
          - human resources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MonthlyPaymentDto'
        '404':
          description: not found
          content:
            application/text*:
              - schema: null
        '409':
          description: conflict
          content:
            application/text:
              - schema: null
    patch:
      tags:
        - monthly-payment-controller
      operationId: updatePayment
      summary: reCalculate payment, Only acces for human resources
      security:
        JWTAuth:
          - human resources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MonthlyPaymentDto'
...
  /monthly-payments/{id}:
    get:
      tags:
        - monthly-payment-controller
      operationId: getPaymentById
      summary: Acces to human resources and personal that request his own information
      security:
        - JWTAuth:
            - human resources
            - employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MonthlyPaymentDto'
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: To acces to this endpoints, its required a valid JWT Token.
  schemas:
    DepartmentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PersonalDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        password:
          type: string
        department:
          $ref: '#/components/schemas/DepartmentDto'
        role:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
        workstation:
          type: string
    RoleDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    PaymentRequest:
      type: object
      properties:
        personalId:
          type: integer
          format: int64
        bonuses:
          type: number
          format: double
...
    MonthlyPaymentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        personalDto:
          $ref: '#/components/schemas/PersonalDto'
        generatedDate:
          type: string
          format: date-time
        paymentMonth:
          type: string
          example: 2025-03
        percentageHealthInsurance:
          type: number
        percentageTaxWithHoldings:
          type: number
        totalPercentageDeductions:
          type: number
        salaryReduction:
          type: number
        Bonuses:
          type: number
        overtimes:
          type: number
        grossSalary:
          type: number
        netSalary:
          type: number